refactored component names
