+++ react/components/PublishThumbnailInput.jsx
