--- react/components/metadataInputs.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class MetadataInputs extends React.Component {
-  render () {
-    return (
-      <div>
-        <h3>details component</h3>
-      </div>
-    );
-  }
-}
-
-module.exports = MetadataInputs;
